+++ backend/src/controllers/user.controller.js
@@ -1,30 +1,39 @@
 const UserModel = require("../models/User.model");
 const bcrypt = require("bcryptjs");
 
-// --- CONJUNTO DE OPERACIONES HECHAS CON LA BASE DE DATOS "users" ---
-// -- BORRAR UN USUARIO :React ---
-module.exports.deleteUser = (req, res, next) => {
-  const id = req.params.id;
+module.exports.deleteUser = async (req, res) => {
+  try {
+    const id = req.params.id;
 
-  UserModel.findByIdAndDelete(id)
-    .then(() => {
-      res.json({ message: "Usuario eliminado correctamente" });
-    })
-    .catch((err) => {
-      res.json(err);
-    });
+    if (req.user._id.toString() !== id) {
+      return res.status(403).json({ message: "No autorizado" });
+    }
+
+    await UserModel.findByIdAndDelete(id);
+    res.json({ message: "Usuario eliminado correctamente" });
+  } catch (err) {
+    res.status(500).json({ message: err.message });
+  }
 };
 
-// -- ACTUALIZAR UN USUARIO :React --
-module.exports.updateUser = (req, res, next) => {
-  const id = req.params.id;
-  const updates = req.body;
+module.exports.updateUser = async (req, res) => {
+  try {
+    const id = req.params.id;
 
-  UserModel.findByIdAndUpdate(id, updates, { new: true })
-    .then((user) => {
-      res.json(user);
-    })
-    .catch((err) => {
-      res.json(err);
-    });
+    if (req.user._id.toString() !== id) {
+      return res.status(403).json({ message: "No autorizado" });
+    }
+
+    const updates = { ...req.body };
+
+    if (updates.password) {
+      const salt = bcrypt.genSaltSync(10);
+      updates.password = bcrypt.hashSync(updates.password, salt);
+    }
+
+    const user = await UserModel.findByIdAndUpdate(id, updates, { new: true });
+    res.json(user);
+  } catch (err) {
+    res.status(500).json({ message: err.message });
+  }
 };
